+++ test/NUnit.Xml.TestLogger.AcceptanceTests/NUnitTestLoggerAcceptanceTests.cs
@@ -47,8 +47,8 @@ namespace NUnit.Xml.TestLogger.AcceptanceTests
             Assert.AreEqual(ExpectedTestCaseCount, node.Attribute(XName.Get("testcasecount")).Value);
             Assert.AreEqual(ExpectedTestCasePassedCount, node.Attribute(XName.Get("passed")).Value);
             Assert.AreEqual("14", node.Attribute(XName.Get("failed")).Value);
-            Assert.AreEqual("8", node.Attribute(XName.Get("inconclusive")).Value);
-            Assert.AreEqual("6", node.Attribute(XName.Get("skipped")).Value);
+            Assert.AreEqual("6", node.Attribute(XName.Get("inconclusive")).Value);
+            Assert.AreEqual("8", node.Attribute(XName.Get("skipped")).Value);
             Assert.AreEqual("Failed", node.Attribute(XName.Get("result")).Value);
 
             // Start time and End time should be valid dates
@@ -65,8 +65,8 @@ namespace NUnit.Xml.TestLogger.AcceptanceTests
             Assert.AreEqual(ExpectedTestCaseCount, node.Attribute(XName.Get("total")).Value);
             Assert.AreEqual(ExpectedTestCasePassedCount, node.Attribute(XName.Get("passed")).Value);
             Assert.AreEqual("14", node.Attribute(XName.Get("failed")).Value);
-            Assert.AreEqual("8", node.Attribute(XName.Get("inconclusive")).Value);
-            Assert.AreEqual("6", node.Attribute(XName.Get("skipped")).Value);
+            Assert.AreEqual("6", node.Attribute(XName.Get("inconclusive")).Value);
+            Assert.AreEqual("8", node.Attribute(XName.Get("skipped")).Value);
             Assert.AreEqual("Failed", node.Attribute(XName.Get("result")).Value);
             Assert.AreEqual("NUnit.Xml.TestLogger.NetCore.Tests.dll", node.Attribute(XName.Get("name")).Value);
             Assert.AreEqual(DotnetTestFixture.TestAssembly, node.Attribute(XName.Get("fullname")).Value);
@@ -93,8 +93,8 @@ namespace NUnit.Xml.TestLogger.AcceptanceTests
             Assert.AreEqual("29", node.Attribute(XName.Get("total")).Value);
             Assert.AreEqual("15", node.Attribute(XName.Get("passed")).Value);
             Assert.AreEqual("7", node.Attribute(XName.Get("failed")).Value);
-            Assert.AreEqual("4", node.Attribute(XName.Get("inconclusive")).Value);
-            Assert.AreEqual("3", node.Attribute(XName.Get("skipped")).Value);
+            Assert.AreEqual("3", node.Attribute(XName.Get("inconclusive")).Value);
+            Assert.AreEqual("4", node.Attribute(XName.Get("skipped")).Value);
             Assert.AreEqual("Failed", node.Attribute(XName.Get("result")).Value);
             Assert.AreEqual("NUnit.Xml.TestLogger.NetFull", node.Attribute(XName.Get("fullname")).Value);
         }
@@ -111,8 +111,8 @@ namespace NUnit.Xml.TestLogger.AcceptanceTests
             Assert.AreEqual("24", node.Attribute(XName.Get("total")).Value);
             Assert.AreEqual("10", node.Attribute(XName.Get("passed")).Value);
             Assert.AreEqual("7", node.Attribute(XName.Get("failed")).Value);
-            Assert.AreEqual("4", node.Attribute(XName.Get("inconclusive")).Value);
-            Assert.AreEqual("3", node.Attribute(XName.Get("skipped")).Value);
+            Assert.AreEqual("3", node.Attribute(XName.Get("inconclusive")).Value);
+            Assert.AreEqual("4", node.Attribute(XName.Get("skipped")).Value);
             Assert.AreEqual("Failed", node.Attribute(XName.Get("result")).Value);
             Assert.AreEqual("NUnit.Xml.TestLogger.Tests2", node.Attribute(XName.Get("fullname")).Value);
         }